refactor(billing-and-subscriptions): Remove legacy gyvatukas system and implement subscription automation with export capabilities
- Remove deprecated gyvatukas calculation system including calculator service, batch processor, and related contracts
- Remove legacy billing calculation services (BillingCalculator, ElectricityCalculator, HeatingCalculator, WaterCalculator)
- Remove gyvatukas-specific database migrations, audit models, and performance monitoring
- Implement SubscriptionAutomationService for handling subscription expiry notifications and auto-renewal
- Add MonitorSubscriptionsCommand for daily subscription monitoring and renewal triggering
- Create SubscriptionRenewal model and Filament resource for managing renewal history
- Implement ScheduledExportService with daily, weekly, and monthly export commands
- Add CleanupExportFiles command for managing exported file lifecycle
- Create AutoRenewalFailedNotification and SuperadminRenewalFailureNotification for renewal alerts
- Add DashboardCacheService for optimized dashboard metric caching
- Implement SubscriptionRenewal database migration with auto-renewal configuration fields
- Add activity logs report view and export functionality
- Update BuildingResource and BuildingController to remove gyvatukas-related operations
- Add property test for export data completeness validation
- Add unit tests for ExportService functionality
- Update validation engine with enhanced consumption and seasonal validators
- Consolidate database indexes and remove legacy calculation column migrations
- This refactoring simplifies the codebase by removing the deprecated gyvatukas system while introducing modern subscription lifecycle management and comprehensive reporting capabilities

+++ app/Models/Organization.php
@@ -320,7 +320,7 @@ class Organization extends Model
             return 0;
         }
 
-        return now()
+        return (int) now()
             ->startOfDay()
             ->diffInDays($this->subscription_ends_at->startOfDay(), false);
     }
